refactor(forgot-password): Add function to handle sending password reset email
